Fixes, breaks
ReVal connection is... broken? Not sure.

Updated to rc.51 to fix some things. Minor config changes as a result.

(To handle local servers, HTTP scheme is now part of the host string.)

Broke some API code into separate files to keep same things with same.
Still in the package, though.

Algo two is working, but needs work on the session id.

Mostly a commit for safety/checkpointing.
