+++ api/api_test.go
@@ -29,7 +29,7 @@ func Test_get_manufactuerer(t *testing.T) {
 
 	for _, tc := range tests {
 		t.Run(fmt.Sprintf("Get Manufactuerer = %v", tc.mfgs), func(t *testing.T) {
-			got := Mac_to_mfg(&cfg, tc.mac)
+			got := MacToMfg(&cfg, tc.mac)
 			if got != tc.mfgs {
 				t.Fatalf("got [ %v ] want [ %v ]", got, tc.mfgs)
 			} else {
@@ -50,7 +50,7 @@ func Test_thrash_db(t *testing.T) {
 
 	for ndx := 0; ndx < dbIterations; ndx++ {
 		t.Run(fmt.Sprintf("Thrash DB = %d", ndx), func(t *testing.T) {
-			got := Mac_to_mfg(cfg, "aa:bb:cc")
+			got := MacToMfg(cfg, "aa:bb:cc")
 			if got != "unknown" {
 				t.Fatalf("got %v; want %v", got, "unknown")
 			} else {
@@ -106,6 +106,7 @@ func Test_StoreContent(t *testing.T) {
 	for _, server := range []string{"directus", "reval"} {
 		svr := config.GetServer(cfg, server)
 		auth, _ := GetToken(svr)
-		StoreDeviceCount(cfg, svr, auth, "Next:0", 1)
+		t.Log("auth", auth)
+		StoreDeviceCount(cfg, svr, auth, 42, "Next:0", 1)
 	}
 }
