+++ api/event.go
@@ -1,0 +1,5 @@
+package api
+
+func LogEvent(id string, info map[string]string) {
+
+}
